WIP - refactor so it's easier to make unit tests
- TODO: move all DAO-calls to cyberService
- Add tests
