+++ src/skills.py
@@ -1,11 +1,10 @@
 from colorama import Fore, Style
 
-import dice, cyberdao as DAO
-from gameHelper import safeCastToInt, printGreenLine, coloredText, list_skills_helper_str, printRedLine, COOL, INT, REF
-from character import Character
-from logger import Log, log_pos, log_neg, log_event, log_neutral
-from roles import roleSpecialAbility, solo, rocker, netrunner, media, nomad, fixer, cop, corp, techie, meditechie
-from skill import SkillInfo
+from src.gameHelper import safeCastToInt, printGreenLine, coloredText, list_skills_helper_str, printRedLine, COOL, INT, REF
+from src.character import Character
+from src.logger import Log, log_pos, log_neg, log_event, log_neutral
+from src.roles import roleSpecialAbility, solo, rocker, netrunner, media, nomad, fixer, cop, corp, techie, meditechie
+from src.skill import SkillInfo
 
 skill_athletics = 'athletics'
 skill_first_aid = 'first aid'
@@ -40,40 +39,6 @@ def printSkillInfo(skills):
     for s in skills:
         skillInfo = skills[s]
         print(f"({s}) {skillInfo['skill']} [{skillInfo['attribute']}]: {skillInfo['description']}")
-
-
-def rollCharacterMeleeDef(name, roll):
-    character = DAO.getCharacterByName(name)
-    if character is not None:
-        skill = DAO.getSkillByName('dodge/escape')
-        if skill is not None:
-            die_roll = safeCastToInt(roll)
-            if die_roll <= 0:
-                (die_roll, _) = dice.rollWithCrit(True)
-
-            atr_bonus = character.attributes[skill['attribute']]
-
-            char_dodge_skill = None
-            for s in character.skills:
-                if s.id == skill['id']:
-                    char_dodge_skill = s
-                    break
-            char_dodge_lvl = 0
-            if char_dodge_skill is not None:
-                char_dodge_lvl = char_dodge_skill.lvl
-            roll = die_roll + atr_bonus + char_dodge_lvl
-            print(f"""(die roll = {die_roll} atr_bonus = {atr_bonus} dodge = {char_dodge_lvl})""")
-            printGreenLine(f"Melee def total: {roll} (hopefully the attacker rolled lower..)")
-
-
-def rollCharacterSkillById(id, skill_num, roll, modifier, added_luck) -> list[Log]:
-    character = DAO.getCharacterById(id)
-    return rollCharacterSkill(character, skill_num, roll, modifier, added_luck)
-
-
-def rollCharacterSkillByName(name, skill_num, roll, modifier):
-    character = DAO.getCharacterByName(name)
-    return rollCharacterSkill(character, skill_num, roll, modifier, added_luck=None)
 
 
 def character_special_atr_bonus_on_skill(character: Character) -> (int, str):
@@ -114,64 +79,6 @@ def character_special_atr_bonus_on_skill(character: Character) -> (int, str):
     return (atr_bonus, atr)
 
 
-#TODO: handle special skill
-def rollCharacterSkill(character, skill_num, roll=0, modifier=0, added_luck=None) -> list[Log]:
-    logs = []
-    skill_name = ''
-    roll_modifier = safeCastToInt(modifier)
-    skill_id = safeCastToInt(skill_num)
-    skill = None
-    if character is not None:
-        t_roll = safeCastToInt(roll)
-        atr_bonus = 0
-        char_skill_lvl = 0
-        die_roll = 0
-        (special_atr_bonus, special_atr) = character_special_atr_bonus_on_skill(character)
-        if skill_id == 0:
-            skill = {
-                'id': 0,
-                'skill': roleSpecialAbility(character.role),
-                'attribute': special_atr,  # TODO: define for all special skills
-                'description': 'TODO'
-            }
-        else:
-            skill = DAO.getSkillById(skill_id)
-
-        if skill is not None:
-            skill_name = skill['skill']
-
-            if t_roll <= 0:
-                if added_luck == None:
-                    added_luck = dice.handleLuck()
-                (die_roll, dice_logs) = dice.rollWithCritAndGivenLuck(added_luck)
-                logs = logs + dice_logs
-            else:
-                die_roll = t_roll
-            skill_with_lvl = None
-            if skill_id == 0:
-                atr_bonus = special_atr_bonus
-                skill_with_lvl = SkillInfo(skill_id, skill['skill'], character.specialAbility, skill['attribute'], is_original=True)
-            else:
-                skill_with_lvl_arr = [s for s in character.skills if s.skill == skill_name]
-                skill_atr = skill['attribute']
-                atr_bonus = character.attributes[skill_atr]
-                if len(skill_with_lvl_arr) > 0:
-                    skill_with_lvl = skill_with_lvl_arr[0]
-                    char_skill_lvl = skill_with_lvl.lvl
-            roll = die_roll + char_skill_lvl + atr_bonus + roll_modifier
-
-            logs = log_event(logs, f"""{character.name} rolled {roll} for {skill_name}""", log_neutral)
-            logs = log_event(
-                logs,
-                f"(die roll = {die_roll} atr_bonus = {atr_bonus} skill_level = {char_skill_lvl} modifier = {roll_modifier})",
-                log_neutral
-            )
-
-        else:
-            logs = log_event(logs, f'SKILL NOT FOUND [skill_id = {skill_id}]', log_neg)
-
-    return logs
-
 def printCharSkillInfo(skills):
     if len(skills) > 0:
         for s in skills:
@@ -179,108 +86,3 @@ def printCharSkillInfo(skills):
             print(f"{s.skill} - {lvl} [{atr}]")
     else:
         print(f'No skills found')
-
-
-def listSkillsByAttribute(atr: str):
-    atr_skills = skillsByAttribute(atr)
-    printSkillInfo(atr_skills)
-
-
-def findSkillsByString(string: str):
-    skills = DAO.skillsByFuzzyLogic(string)
-    printSkillInfo(skills)
-
-
-def awarenessSkill():
-    return DAO.skillByName('awareness')
-
-
-def fetchAllSkils():
-    skills = allSkills().values()
-    return list(skills)
-
-def listAllSkills():
-    all_skills = allSkills()
-    printSkillInfo(all_skills)
-
-def udpateCharacterSkill(character, skill_id, lvl_up_amount) -> list[Log]:
-    event_logs = []
-    t_skill = safeCastToInt(skill_id)
-    if t_skill >= 0:
-        if character is not None:
-            if t_skill == 0:
-                DAO.updateCharSpecial(character.id, character.role, lvl_up_amount)
-                special_log = Log(f'{character.name} special updated (+{lvl_up_amount})', log_pos)
-                event_logs.append(special_log.toJson())
-            else:
-                skill = DAO.getSkillById(skill_id)
-                char_skills: list[SkillInfo] = character.skills
-                curr_skill_lvl = 0
-                for s in char_skills:
-                    if s.id == skill_id:
-                        curr_skill_lvl = s.lvl
-                can_update = curr_skill_lvl + lvl_up_amount <= 10
-                if skill is not None:
-                    if can_update:
-                        DAO.updateCharSkill(character.id, skill, lvl_up_amount)
-                        skill_updated_log = Log(f"Skill {skill['skill']} (+{lvl_up_amount}) updated for {character.name}", log_pos)
-                        event_logs.append(skill_updated_log.toJson())
-                    else:
-                        event_logs.append(Log("Skill can't get above 10", log_neg).toJson())
-                else:
-                    not_found_log = Log(f'Skill not found by id ({skill_id})', log_neg)
-                    event_logs.append(not_found_log.toJson())
-    else:
-        not_valid_skill_log = Log(f"'{skill_id}' not a valid skill id", log_neg)
-        event_logs.append(not_valid_skill_log.toJson())
-
-    return event_logs
-
-
-def updateCharSkillById(char_id, skill_id, lvl_up_amount):
-    character = DAO.getCharacterById(char_id)
-    return udpateCharacterSkill(character, skill_id, lvl_up_amount)
-
-
-def updateCharSkill(name, skill_id, lvl_up_amount):
-    character = DAO.getCharacterByName(name)
-    return udpateCharacterSkill(character, skill_id, lvl_up_amount)
-
-
-def printCharacterSkills(name):
-    skills = characterSkills(name)
-    printCharSkillInfo(skills)
-
-
-def characterSkills(name):
-    character = DAO.getCharacterByName(name)
-    if character is not None:
-        skills = DAO.getCharacterSkillsById(character.id)
-        return skills
-    else:
-        print(f'{name} Not found')
-        return list()
-
-
-def allSkills():
-    skills = DAO.listSkills()
-    return skills
-
-
-def skillsByAttribute(atr):
-    skills = DAO.listSkillsByAttribute(atr)
-    return skills
-
-
-def listSkills(command):
-    match command:
-        case[_]:
-            listAllSkills()
-        case[_, 'atr', atr]:
-            listSkillsByAttribute(atr)
-        case[_, 'fuzzy', str]:
-            findSkillsByString(str)
-        case[_, 'char', name]:
-            printCharacterSkills(name)
-        case _:
-            print(list_skills_helper_str)
